made draw_brushstroke work with transparent images
draw_brushstroke works most of the time but will sometimes return the error: (-209:Sizes of input arguments do not match) The operation is neither 'array op array' (where arrays have the same size and the same number of channels), nor 'array op scalar', nor 'scalar op array' in function 'arithm_op'

It says the number of channels do not match but brushstroke and inImg should both have four channels (fourth channel is for transparency). So I'm not sure why this is happening.